sudoku_engine

Function assign

pub fn assign(
    board: &mut Board,
    idx: usize,
    value: usize,
) -> Result<Elimination, SudokuErrors>

Place the digit value in the puzzle at location idx.

Errors

This function can return an error if

idx is out of bounds for the grid.

value is not a valid digit for the puzzle.